Sommelier: Start to tidy up by splitting into functional packages
