+++ src/main/scala/duesoldi/blog/routes/BlogEditingRoutes.scala
@@ -6,7 +6,7 @@ import duesoldi.blog.validation.ValidateIdentifier
 import duesoldi.blog.{EntriesFromYaml, EntriesToYaml, EntryFromYaml, EntryToYaml}
 import duesoldi.config.Config
 import duesoldi.dependencies.DueSoldiDependencies._
-import sommelier.Controller
+import sommelier.routing.Controller
 
 import scala.concurrent.ExecutionContext
 
@@ -14,8 +14,8 @@ class BlogEditingController(implicit executionContext: ExecutionContext, appConf
 extends Controller
 {
   import duesoldi.app.TempSommelierIntegration._
-  import sommelier.Routing._
-  import sommelier.Unpacking._
+  import sommelier.routing.Routing._
+  import sommelier.handling.Unpacking._
 
   PUT("/admin/blog/:id").Authorization(basicAdminAuth) ->- { implicit context =>
     for {
