+++ src/main/scala/sommelier/routing/Controller.scala
@@ -1,7 +1,8 @@
-package sommelier
+package sommelier.routing
 
 import sommelier.Middleware.{Incoming, IncomingHandler, Outgoing, OutgoingHandler}
-import sommelier.Routing.Handler
+import sommelier.routing.Routing.Handler
+import sommelier.{Middleware, Route}
 
 import scala.collection.mutable
